+++ otx/algorithms/detection/configs/detection/resnet50_deformable_detr/model.py
@@ -1,4 +1,7 @@
 """Model config for Deformable DETR."""
+_base_ = [
+    "../../../../../recipes/stages/detection/incremental.py",
+]
 model = dict(
     type="CustomDeformableDETR",
     backbone=dict(
@@ -76,34 +79,14 @@ model = dict(
 )
 # optimizer
 optimizer = dict(
+    _delete_=True,
     type="AdamW",
     lr=2e-4,
     weight_decay=0.0001,
-    paramwise_cfg=dict(
-        custom_keys={
-            "backbone": dict(lr_mult=0.1),
-            "sampling_offsets": dict(lr_mult=0.1),
-            "reference_points": dict(lr_mult=0.1),
-        }
-    ),
 )
 optimizer_config = dict(grad_clip=dict(max_norm=0.1, norm_type=2))
-# learning policy
-lr_config = dict(policy="step", step=[10])
-runner = dict(type="EpochRunnerWithCancel", max_epochs=12)
 load_from = "https://download.openmmlab.com/mmdetection/v2.0/deformable_detr/\
 deformable_detr_twostage_refine_r50_16x2_50e_coco/\
 deformable_detr_twostage_refine_r50_16x2_50e_coco_20210419_220613-9d28ab72.pth"
 resume_from = None
-
-checkpoint_config = dict(interval=1)
-# yapf:disable
-log_config = dict(
-    interval=100,
-    hooks=[
-        dict(type="TextLoggerHook"),
-    ],
-)
-log_level = "INFO"
-workflow = [("train", 1)]
-task_adapt = dict(op="REPLACE", type="temp", efficient_mode=False, use_mpa_anchor=False)
+ignore = False