+++ otx/cli/tools/eval.py
@@ -18,6 +18,7 @@ import json
 from pathlib import Path
 
 from otx.api.entities.inference_parameters import InferenceParameters
+from otx.api.entities.model_template import TaskType
 from otx.api.entities.resultset import ResultSetEntity
 from otx.api.entities.subset import Subset
 from otx.api.entities.task_environment import TaskEnvironment
@@ -133,7 +134,10 @@ def main():
 
     validation_dataset = dataset.get_subset(Subset.TESTING)
     predicted_validation_dataset = task.infer(
-        validation_dataset.with_empty_annotations(),
+        # temp (sungchul): remain annotation for visual prompting
+        validation_dataset
+        if getattr(task, "task_type", None) == TaskType.VISUAL_PROMPTING
+        else validation_dataset.with_empty_annotations(),
         InferenceParameters(is_evaluation=False),
     )