+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Add per-class XAI saliency maps for Mask R-CNN model (https://github.com/openvinotoolkit/training_extensions/pull/2227)
 - Add new object detector Deformable DETR (<https://github.com/openvinotoolkit/training_extensions/pull/2249>)
 - Add new object detecotr DINO(<https://github.com/openvinotoolkit/training_extensions/pull/2266>)
+- Add new visual prompting task (https://github.com/openvinotoolkit/training_extensions/pull/2203)
 
 ### Enhancements